m1 - add resource manager for tcp
